Handle inputs of several args or mixed

+++ push_swap.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   push_swap.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tjorge-l <tjorge-l@student.42.fr>          +#+  +:+       +#+        */
+/*   By: tjorge-l <tjorge-l@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/24 17:38:19 by tjorge-l          #+#    #+#             */
-/*   Updated: 2024/06/03 16:01:00 by tjorge-l         ###   ########.fr       */
+/*   Updated: 2024/06/04 10:06:08 by tjorge-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ typedef struct stack {
 
 void	ft_error(void);
 
-long	ft_atoi_alt(char *str);
+long long	ft_atoi_alt(char *str);
 
 void	free_stack(stack **s);
 stack	*last_on_stack(stack **s);
